// DXC_Frontend/src/api/models/ticketDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface TicketDto {
  publicId?: string
  /** @nullable */
  name?: string | null
  /** @nullable */
  description?: string | null
  price?: number
  childPrice?: number
  /** @nullable */
  priceCurrency?: string | null
  thuTu?: number
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
}